code/agent/chatflow/requirements.txt
# 大模型相关
openai>=1.0.0
transformers>=4.40.0
torch>=2.0.0
sentence-transformers>=2.5.0
accelerate>=0.27.0

# 向量数据库
chromadb>=0.5.0
faiss-cpu>=1.7.4

# RAG重排序
FlagEmbedding>=1.2.0

# 工具调用
pydantic>=2.0.0
jinja2>=3.1.0

# 网页界面
streamlit>=1.30.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 搜索工具
requests>=2.31.0
beautifulsoup4>=4.12.0

# 配置管理
pyyaml>=6.0

# 评测和可视化
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# 异步处理
aiohttp>=3.9.0
asyncio>=3.4.0

# 文本处理
jieba>=0.42.1
